Making ARMWidenStrings to be target independent
