+++ llvm/include/llvm/Analysis/TargetTransformInfoImpl.h
@@ -1006,6 +1006,8 @@ public:
 
   unsigned getMaxNumArgs() const { return UINT_MAX; }
 
+  bool useWidenGlobalStrings() const { return false; }
+
 protected:
   // Obtain the minimum required size to hold the value (without the sign)
   // In case of a vector it returns the min required size for one element.
