--- llvm/include/llvm/Transforms/Scalar/ARMWidenStrings.h
@@ -1,30 +1,0 @@
-//===- ARMWidenStrings.h --------------------------------------*- C++ -*-===//
-//
-// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-//
-//===----------------------------------------------------------------------===//
-//
-// This file provides the interface for the ArmWidenStrings pass
-//
-//===----------------------------------------------------------------------===//
-
-#ifndef LLVM_TRANSFORMS_SCALAR_ARMWIDENSTRINGS_H
-#define LLVM_TRANSFORMS_SCALAR_ARMWIDENSTRINGS_H
-
-#include "llvm/IR/PassManager.h"
-
-namespace llvm {
-
-class Module;
-
-class ARMWidenStringsPass : public PassInfoMixin<ARMWidenStringsPass> {
-public:
-  ARMWidenStringsPass() = default;
-  PreservedAnalyses run(Function &F, FunctionAnalysisManager &);
-};
-
-} // end namespace llvm
-
-#endif // LLVM_TRANSFORMS_SCALAR_ARMWIDENSTRINGS_H
